fixed network: size_t instead of forcing the type

+++ include/GEngine/net/net_server.hpp
@@ -35,10 +35,10 @@ public:
         return m_socketv6;
     }
 
-    size_t getNumClients(void) const {
+    uint32_t getNumClients(void) const {
         return m_clients.size();
     }
-    size_t getMaxClients(void) const {
+    uint32_t getMaxClients(void) const {
         return m_maxClients;
     };
